Move DateTime2ToCommonDateFormats to implicits.Dates trait
