+++ common/app/model/Cached.scala
@@ -5,7 +5,7 @@ import org.joda.time.DateTime
 import org.scala_tools.time.Imports._
 import play.api.mvc.{ SimpleResult, Results }
 
-object Cached extends Results {
+object Cached extends Results with implicits.Dates {
 
   def apply(seconds: Int)(result: SimpleResult): SimpleResult = {
     if (result.header.status == 200) cacheHeaders(seconds, result) else result
